feat(*): upgrade to payload v2

+++ src/api/payload-crowdin-sync/files.ts
@@ -16,6 +16,8 @@ import {
 } from "../helpers";
 import { isEmpty } from "lodash";
 
+import { CrowdinArticleDirectory } from '../../payload-types'
+
 export interface IcrowdinFile {
   id: string;
   originalId: number;
@@ -125,7 +127,7 @@ export class payloadCrowdinSyncFilesApi {
       const crowdinDirectory = await this.sourceFilesApi.createDirectory(
         this.projectId,
         {
-          directoryId: crowdinPayloadCollectionDirectory.originalId,
+          directoryId: crowdinPayloadCollectionDirectory.originalId as number,
           name: global ? collectionSlug : document.id,
           title: global
             ? toWords(collectionSlug)
@@ -419,9 +421,9 @@ export class payloadCrowdinSyncFilesApi {
     }
   }
 
-  async getArticleDirectory(documentId: string) {
+  async getArticleDirectory(documentId: string): Promise<CrowdinArticleDirectory | undefined> {
     const result = await getArticleDirectory(documentId, this.payload);
-    return result;
+    return result as CrowdinArticleDirectory | undefined;
   }
 
   async deleteFilesAndDirectory(documentId: string) {
@@ -438,6 +440,9 @@ export class payloadCrowdinSyncFilesApi {
     const crowdinPayloadArticleDirectory = await this.getArticleDirectory(
       documentId
     );
+    if (!crowdinPayloadArticleDirectory || !crowdinPayloadArticleDirectory.originalId) {
+      return
+    }
     await this.sourceFilesApi.deleteDirectory(
       this.projectId,
       crowdinPayloadArticleDirectory.originalId
